Flytte vedtaksmeny i egen fil
